+++ src/client/java/com/roulettepaymenttracker/client/PlayerDataHolder.java
@@ -1,3 +1,3 @@
 package com.roulettepaymenttracker.client;
 
-public record PlayerDataHolder(String username, int amount) {}
+public record PlayerDataHolder(String username, long amount) {}
